New invoker, TreeViewInvoker
This new invoker is able to handle all the palettes of one treeview,
removing the need of one invoker per cell
renderer (CellRendererInvoker).  This simplifies the code and makes it
more efficient.  Also removes the logic from the CellRendererIcon, which
should only care about drawing itself.

Note this is work in progress.  Click and long press do not raise the
palette yet.  Only mouse over at the moment.

This commit also comments out the invoker of CellRendererIcon.

FIXME, explain API.