Add spent-times to activity metadata
spent-times is a comma-separated list with the times the user has spent
on a particular activity instance.  Each time the instance is resumed,
the activity duration is added to the list, measured in seconds.  The
time the activity is in the background is discriminated (ie. doesn't
added).  And internally, microseconds are used for the computation.
Example of an activity instance that has been resumed four times:

    '110, 35, 12, 143'